+++ DelvUI/Interface/HudHelper.cs
@@ -1,139 +1,38 @@
-/*
-Copyright(c) 2021 jkcclemens HUD Manager
-Modifications Copyright(c) 2021 DelvUI
-09/27/2021 - Extracted code to move in-game UI elements.
-
-This program is free software: you can redistribute it and/or modify
-it under the terms of the GNU General Public License as published by
-the Free Software Foundation, either version 3 of the License, or
-(at your option) any later version.
-
-This program is distributed in the hope that it will be useful,
-but WITHOUT ANY WARRANTY; without even the implied warranty of
-MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-GNU General Public License for more details.
-
-You should have received a copy of the GNU General Public License
-along with this program.  If not, see <http://www.gnu.org/licenses/>.
-*/
-
+using Dalamud.Game.Addon.Lifecycle;
 using Dalamud.Game.ClientState.Conditions;
-using Dalamud.Game.ClientState.Objects.Enums;
 using Dalamud.Game.ClientState.Objects.SubKinds;
-using Dalamud.Game.ClientState.Objects.Types;
-using Dalamud.Logging;
 using DelvUI.Config;
 using DelvUI.Helpers;
 using DelvUI.Interface.GeneralElements;
-using DelvUI.Interface.Jobs;
 using FFXIVClientStructs.FFXIV.Component.GUI;
+using ImGuiNET;
 using System;
 using System.Collections.Generic;
-using System.Linq;
 using System.Numerics;
-using System.Runtime.CompilerServices;
-using System.Runtime.InteropServices;
 
 namespace DelvUI.Interface
 {
     public class HudHelper : IDisposable
     {
-        private delegate void SetPositionDelegate(IntPtr addon, short x, short y);
-        private delegate IntPtr GetBaseUIObjectDelegate();
-        private delegate byte UpdateAddonPositionDelegate(IntPtr manager, IntPtr addon, byte clicked);
-        private delegate IntPtr GetFilePointerDelegate(byte index);
-
         private HUDOptionsConfig Config => ConfigurationManager.Instance.GetConfigObject<HUDOptionsConfig>();
-
-        private readonly uint[] _goldSaucerIDs = { 144, 388, 389, 390, 391, 579, 792, 899, 941 };
-
-        private GetBaseUIObjectDelegate? _getBaseUIObject;
-        private SetPositionDelegate? _setPosition;
-        private UpdateAddonPositionDelegate? _updateAddonPosition;
-        private GetFilePointerDelegate? _getFilePointer = null;
 
         private readonly string[] _hotbarAddonNames = { "_ActionBar", "_ActionBar01", "_ActionBar02", "_ActionBar03", "_ActionBar04", "_ActionBar05", "_ActionBar06", "_ActionBar07", "_ActionBar08", "_ActionBar09" };
         private bool? _previousHotbarCrossVisible = null;
 
+        private Vector2 _castBarPos = Vector2.Zero;
+        private bool _hidingCastBar = false;
+        private Dictionary<string, Vector2> _jobGaugePos = new();
+        private bool _hidingJobGauge = false;
+        private Vector2 _pullTimerPos = Vector2.Zero;
+        private bool _hidingPullTimer = false;
+
+        private uint _previousJob = 0;
+        private double _jobChangeTime = 0;
+        private bool _jobChangeUpdated = false;
+
         public HudHelper()
         {
-            #region Signatures
             Config.ValueChangeEvent += ConfigValueChanged;
-
-            /*
-             Part of getBaseUiObject disassembly signature
-            .text:00007FF6481C2F60                   Component__GUI__AtkStage_GetSingleton1 proc near
-            .text:00007FF6481C2F60 48 8B 05 99 04 8D+mov     rax, cs:g_AtkStage
-            .text:00007FF6481C2F60 01
-            .text:00007FF6481C2F67 C3                retn
-            .text:00007FF6481C2F67                   Component__GUI__AtkStage_GetSingleton1 endp
-            .text:00007FF6481C2F67
-            */
-            IntPtr getBaseUiObjectPtr = Plugin.SigScanner.ScanText("E8 ?? ?? ?? ?? 0F BF D5");
-            _getBaseUIObject = Marshal.GetDelegateForFunctionPointer<GetBaseUIObjectDelegate>(getBaseUiObjectPtr);
-
-            /*
-             Part of setPosition disassembly signature
-            .text:00007FF6481BFF20                   Component__GUI__AtkUnitBase_SetPosition proc near
-            .text:00007FF6481BFF20 4C 8B 89 C8 00 00+mov     r9, [rcx+0C8h]
-            .text:00007FF6481BFF20 00
-            .text:00007FF6481BFF27 41 0F BF C0       movsx   eax, r8w
-            .text:00007FF6481BFF2B 66 89 91 BC 01 00+mov     [rcx+1BCh], dx
-            .text:00007FF6481BFF2B 00
-            .text:00007FF6481BFF32 66 44 89 81 BE 01+mov     [rcx+1BEh], r8w
-            .text:00007FF6481BFF32 00 00
-            .text:00007FF6481BFF3A 66 0F 6E C8       movd    xmm1, eax
-            .text:00007FF6481BFF3E 0F BF C2          movsx   eax, dx
-            .text:00007FF6481BFF41 0F 5B C9          cvtdq2ps xmm1, xmm1
-            .text:00007FF6481BFF44 66 0F 6E D0       movd    xmm2, eax
-            .text:00007FF6481BFF48 0F 5B D2          cvtdq2ps xmm2, xmm2
-            .text:00007FF6481BFF4B 4D 85 C9          test    r9, r9
-            .text:00007FF6481BFF4E 74 3B             jz      short locret_7FF6481BFF8B
-            */
-            IntPtr setPositionPtr = Plugin.SigScanner.ScanText("4C 8B 89 ?? ?? ?? ?? 41 0F BF C0");
-            _setPosition = Marshal.GetDelegateForFunctionPointer<SetPositionDelegate>(setPositionPtr);
-
-            /*
-            Part of updateAddonPosition disassembly signature
-            .text:00007FF6481CF020                   sub_7FF6481CF020 proc near
-            .text:00007FF6481CF020
-            .text:00007FF6481CF020                   arg_0= qword ptr  8
-            .text:00007FF6481CF020
-            .text:00007FF6481CF020 48 89 5C 24 08    mov     [rsp+arg_0], rbx
-            .text:00007FF6481CF025 57                push    rdi
-            .text:00007FF6481CF026 48 83 EC 20       sub     rsp, 20h
-            .text:00007FF6481CF02A 48 8B DA          mov     rbx, rdx
-            .text:00007FF6481CF02D 48 8B F9          mov     rdi, rcx
-            .text:00007FF6481CF030 48 85 D2          test    rdx, rdx
-            .text:00007FF6481CF033 0F 84 CA 00 00 00 jz      loc_7FF6481CF103
-            */
-            IntPtr updateAddonPositionPtr = Plugin.SigScanner.ScanText("E8 ?? ?? ?? ?? 48 8B 8B ?? ?? ?? ?? 33 D2 48 8B 01 FF 90 ?? ?? ?? ??");
-            _updateAddonPosition = Marshal.GetDelegateForFunctionPointer<UpdateAddonPositionDelegate>(updateAddonPositionPtr);
-
-            var getFilePointerPtr = Plugin.SigScanner.ScanText("E8 ?? ?? ?? ?? 48 85 C0 74 14 83 7B 44 00");
-            if (getFilePointerPtr != IntPtr.Zero)
-            {
-                _getFilePointer = Marshal.GetDelegateForFunctionPointer<GetFilePointerDelegate>(getFilePointerPtr);
-            }
-            #endregion
-        }
-
-        internal static byte GetStatus(GameObject actor)
-        {
-            // 40 57 48 83 EC 70 48 8B F9 E8 ?? ?? ?? ?? 81 BF ?? ?? ?? ?? ?? ?? ?? ??
-            const int offset = 0x19A0;
-            return Marshal.ReadByte(actor.Address + offset);
-        }
-
-        internal int GetActiveHUDLayoutIndex()
-        {
-            if (_getFilePointer == null) { return 0; }
-
-            IntPtr dataPtr = _getFilePointer.Invoke(0) + 0x50;
-            IntPtr slotPtr = Marshal.ReadIntPtr(dataPtr) + 0x59e8;
-            int index = Marshal.ReadInt32(slotPtr);
-
-            return Math.Clamp(index, 0, 3);
         }
 
         ~HudHelper()
@@ -165,9 +64,33 @@ namespace DelvUI.Interface
         public void Update()
         {
             UpdateCombatActionBars();
-            UpdateJobGauges();
-            UpdateDefaultCastBar();
-            UpdateDefaultPulltimer();
+
+            // detect job change
+            PlayerCharacter? player = Plugin.ClientState.LocalPlayer;
+            if (player != null)
+            {
+                uint job = player.ClassJob.Id;
+                double now = ImGui.GetTime();
+
+                if (job != _previousJob)
+                {
+                    _hidingCastBar = false;
+                    UpdateDefaultCastBar();
+
+                    _hidingPullTimer = false;
+                    UpdateDefaultPulltimer();
+
+                    _previousJob = job;
+                    _jobChangeUpdated = false;
+                    _jobChangeTime = now;
+                }
+                else if (!_jobChangeUpdated && now - _jobChangeTime > 0.01f)
+                {
+                    _jobChangeUpdated = true;
+                    _hidingJobGauge = false;
+                    UpdateJobGauges();
+                }
+            }
         }
 
         public bool IsElementHidden(HudElement element)
@@ -189,7 +112,6 @@ namespace DelvUI.Interface
                     UpdateDefaultPulltimer();
                     break;
                 case "HideDefaultJobGauges":
-                case "DisableJobGaugeSounds":
                     UpdateJobGauges();
                     break;
             }
@@ -197,9 +119,19 @@ namespace DelvUI.Interface
 
         private void UpdateCombatActionBars()
         {
-            if (Plugin.Condition[ConditionFlag.OccupiedInEvent] || 
+            if (Plugin.Condition[ConditionFlag.OccupiedInEvent] ||
                 Plugin.Condition[ConditionFlag.OccupiedInQuestEvent] ||
                 Plugin.Condition[ConditionFlag.ChocoboRacing] ||
+                Plugin.Condition[ConditionFlag.Occupied] ||
+                Plugin.Condition[ConditionFlag.Occupied30] ||
+                Plugin.Condition[ConditionFlag.Occupied33] ||
+                Plugin.Condition[ConditionFlag.Occupied38] ||
+                Plugin.Condition[ConditionFlag.Occupied39] ||
+                Plugin.Condition[ConditionFlag.OccupiedInCutSceneEvent] ||
+                Plugin.Condition[ConditionFlag.OccupiedInEvent] ||
+                Plugin.Condition[ConditionFlag.OccupiedInQuestEvent] ||
+                Plugin.Condition[ConditionFlag.WatchingCutscene] ||
+                Plugin.Condition[ConditionFlag.WatchingCutscene78] ||
                 Plugin.ClientState.IsPvP)
             {
                 return;
@@ -244,150 +176,131 @@ namespace DelvUI.Interface
             ChatHelper.SendChatMessage("/crosshotbardisplay" + " " + onOffText);
         }
 
-        private void SetAddonVisible(IntPtr addon, bool visible, Vector2 originalPosition)
-        {
-            if (_getBaseUIObject == null || _setPosition == null || _updateAddonPosition == null) { return; }
-
-            IntPtr baseUi = _getBaseUIObject();
-            IntPtr manager = Marshal.ReadIntPtr(baseUi + 0x20);
-
-            short x = visible ? (short)originalPosition.X : (short)32000;
-            short y = visible ? (short)originalPosition.Y : (short)32000;
-
-            _updateAddonPosition(manager, addon, 1);
-            _setPosition(addon, x, y);
-            _updateAddonPosition(manager, addon, 0);
-        }
-
-        private unsafe bool UpdateAddonOriginalPosition(AtkUnitBase* addon, ref Vector2 originalPosition)
-        {
-            if (addon == null) { return false; }
-
-            const float outOfScreen = 30000f;
-            Vector2 addonPosition = new(addon->X, addon->Y);
-
-            if (addonPosition != Vector2.Zero && addonPosition.X < outOfScreen && addonPosition.Y < outOfScreen)
-            {
-                originalPosition = addonPosition;
-            }
-
-            // in case something goes wrong, restore position to 0 so the element is not permanently lost off screen
-            if (originalPosition.X > outOfScreen || originalPosition.Y > outOfScreen)
-            {
-                originalPosition = Vector2.Zero;
-            }
-
-            return addonPosition.X < outOfScreen || addonPosition.Y < outOfScreen;
-        }
-
         private unsafe void UpdateDefaultCastBar(bool forceVisible = false)
         {
-            AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("_CastBar", 1);
-            if (addon == null) { return; }
-
-            int index = GetActiveHUDLayoutIndex();
-            Vector2 previousPos = Config.CastBarOriginalPositions[index];
-            bool isVisible = UpdateAddonOriginalPosition(addon, ref Config.CastBarOriginalPositions[index]);
-
-            if (previousPos != Config.CastBarOriginalPositions[index])
+            if (Config.HideDefaultCastbar && !_hidingCastBar)
             {
-                ConfigurationManager.Instance.SaveConfigurations(true);
+                AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("_CastBar", 1);
+                if (addon != null)
+                {
+                    _castBarPos = new Vector2(addon->RootNode->GetX(), addon->RootNode->GetY());
+                    addon->RootNode->SetPositionFloat(-9999, -9999);
+                }
+
+                _hidingCastBar = true;
+            }
+            else if ((forceVisible || !Config.HideDefaultCastbar) && _hidingCastBar)
+            {
+                AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("_CastBar", 1);
+                if (addon != null)
+                {
+                    addon->RootNode->SetPositionFloat(_castBarPos.X, _castBarPos.Y);
+                }
+
+                _hidingCastBar = false;
             }
 
-            if (isVisible != Config.HideDefaultCastbar && !forceVisible) { return; }
-
-            SetAddonVisible((IntPtr)addon, forceVisible || !Config.HideDefaultCastbar, Config.CastBarOriginalPositions[index]);
+            return;
         }
 
         private unsafe void UpdateDefaultPulltimer(bool forceVisible = false)
         {
-            AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("ScreenInfo_CountDown", 1);
-            if (addon == null) { return; }
-
-            int index = GetActiveHUDLayoutIndex();
-            Vector2 previousPos = Config.PulltimerOriginalPositions[index];
-            bool isVisible = UpdateAddonOriginalPosition(addon, ref Config.PulltimerOriginalPositions[index]);
-
-            if (previousPos != Config.PulltimerOriginalPositions[index])
+            if (Config.HideDefaultPulltimer && !_hidingPullTimer)
             {
-                ConfigurationManager.Instance.SaveConfigurations(true);
+                Plugin.AddonLifecycle.RegisterListener(AddonEvent.PreDraw, "ScreenInfo_CountDown", (addonEvent, args) =>
+                {
+                    AtkUnitBase* addon = (AtkUnitBase*)args.Addon;
+
+                    if (!_hidingPullTimer)
+                    {
+                        _pullTimerPos = new Vector2(addon->RootNode->GetX(), addon->RootNode->GetY());
+                    }
+
+                    addon->RootNode->SetPositionFloat(-9999.0f, -9999.0f);
+                });
+
+                AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("ScreenInfo_CountDown", 1);
+                if (addon != null)
+                {
+                    _pullTimerPos = new Vector2(addon->RootNode->GetX(), addon->RootNode->GetY());
+                    addon->RootNode->SetPositionFloat(-9999, -9999);
+                }
+
+                _hidingPullTimer = true;
+            }
+            else if ((forceVisible || !Config.HideDefaultPulltimer) && _hidingPullTimer)
+            {
+                Plugin.AddonLifecycle.UnregisterListener(AddonEvent.PreDraw, "ScreenInfo_CountDown");
+
+                AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName("ScreenInfo_CountDown", 1);
+                if (addon != null)
+                {
+                    addon->RootNode->SetPositionFloat(_pullTimerPos.X, _pullTimerPos.Y);
+                }
+
+                _hidingPullTimer = false;
             }
 
-            if (isVisible != Config.HideDefaultPulltimer && !forceVisible) { return; }
-
-            SetAddonVisible((IntPtr)addon, forceVisible || !Config.HideDefaultPulltimer, Config.PulltimerOriginalPositions[index]);
+            return;
         }
 
         private unsafe void UpdateJobGauges(bool forceVisible = false)
         {
-            var (addons, names) = FindAddonsStartingWith("JobHud");
+            PlayerCharacter? player = Plugin.ClientState.LocalPlayer;
+            if (player == null) { return; }
 
-            int index = GetActiveHUDLayoutIndex();
-            Dictionary<string, Vector2> dict = Config.JobGaugeOriginalPositions[index];
+            string jobName = JobsHelper.JobNames[player.ClassJob.Id];
 
-            for (int i = 0; i < addons.Count; i++)
+            if (Config.HideDefaultJobGauges && !_hidingJobGauge)
             {
-                AtkUnitBase* addon = (AtkUnitBase*)addons[i];
-                string name = names[i];
+                int i = 0;
+                bool stop = false;
 
-                bool existed = dict.TryGetValue(name, out Vector2 pos);
-
-                Vector2 previousPos = pos;
-                UpdateAddonOriginalPosition(addon, ref pos);
-
-                if (previousPos != pos || !existed)
+                do
                 {
-                    dict[name] = pos;
-                    ConfigurationManager.Instance.SaveConfigurations(true);
-                }
+                    string addonName = $"JobHud{jobName}{i}";
 
-                SetAddonVisible((IntPtr)addon, forceVisible || !Config.HideDefaultJobGauges, dict[name]);
+                    AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName(addonName, 1);
+                    if (addon == null)
+                    {
+                        stop = true;
+                    }
+                    else
+                    {
+                        _jobGaugePos[addonName] = new Vector2(addon->RootNode->GetX(), addon->RootNode->GetY());
+                        addon->RootNode->SetPositionFloat(-9999, -9999);
+                    }
+
+                    i++;
+                } while (!stop);
+
+                _hidingJobGauge = true;
+            }
+            else if ((forceVisible || !Config.HideDefaultJobGauges) && _hidingJobGauge)
+            {
+                int i = 0;
+                bool stop = false;
+
+                do
+                {
+                    string addonName = $"JobHud{jobName}{i}";
+
+                    AtkUnitBase* addon = (AtkUnitBase*)Plugin.GameGui.GetAddonByName(addonName, 1);
+                    if (addon == null)
+                    {
+                        stop = true;
+                    }
+                    else if (_jobGaugePos.TryGetValue(addonName, out Vector2 pos))
+                    {
+                        addon->RootNode->SetPositionFloat(pos.X, pos.Y);
+                    }
+
+                    i++;
+                } while (!stop);
+
+                _hidingJobGauge = false;
             }
         }
-
-        public static unsafe (List<IntPtr>, List<string>) FindAddonsStartingWith(string startingWith)
-        {
-            List<IntPtr> addons = new();
-            List<string> names = new();
-
-            AtkStage* stage = AtkStage.GetSingleton();
-            if (stage == null)
-            {
-                return (addons, names);
-            }
-
-            AtkUnitList* loadedUnitsList = &stage->RaptureAtkUnitManager->AtkUnitManager.AllLoadedUnitsList;
-            if (loadedUnitsList == null) { return (addons, names); }
-
-            for (int i = 0; i < loadedUnitsList->Count; i++)
-            {
-                try
-                {
-                    AtkUnitBase* addon = *(AtkUnitBase**)Unsafe.AsPointer(ref loadedUnitsList->EntriesSpan[i]);
-                    if (addon == null) { continue; }
-
-                    string? name = Marshal.PtrToStringAnsi(new IntPtr(addon->Name));
-                    if (name == null || !name.StartsWith(startingWith)) { continue; }
-
-                    addons.Add((IntPtr)addon);
-                    names.Add(name);
-                }
-                catch { }
-            }
-
-            return (addons, names);
-        }
-        #region Helpers
-
-        private bool IsInCombat() => Plugin.Condition[ConditionFlag.InCombat];
-
-        private bool IsInDuty() => Plugin.Condition[ConditionFlag.BoundByDuty];
-
-        private bool IsCrafting() => Plugin.Condition[ConditionFlag.Crafting];
-
-        private bool HasWeaponDrawn() => (Plugin.ClientState.LocalPlayer != null && Plugin.ClientState.LocalPlayer.StatusFlags.HasFlag(StatusFlags.WeaponOut));
-
-        #endregion
     }
 
     internal class StrataLevelComparer<TKey> : IComparer<TKey> where TKey : PluginConfigObject